input two numbers

+++ Print_the_quotient_of_the_two_numbers_with_the_decimal_point.py
@@ -1,5 +1,5 @@
 #input two numbers
-num1 = int(input("Input 1st number: "))
-num2 = int(input("Input 2nd number: "))
+num1 = float(input("Input 1st number: "))
+num2 = float(input("Input 2nd number: "))
 
-#Pirint quotient
+#Print quotient
